+++ example/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
-import { SinWave, Blink, TerribleNames } from 'useless-things'
+import { SinWave, Blink, TerribleName } from 'useless-things'
 
 export default class App extends React.Component {
   render () {
@@ -15,9 +15,9 @@ export default class App extends React.Component {
 
         <Text style={styles.thing}>Terrible Name Generator</Text>
         <Text style={styles.text}>
-          {TerribleNames() + '\n'}
-          {TerribleNames() + '\n'}
-          {TerribleNames() + '\n'}
+          Random 1: {TerribleName() + '\n'}
+          Random 2: {TerribleName() + '\n'}
+          Random 3: {TerribleName() + '\n'}
         </Text>
       </View>
     );
@@ -42,6 +42,6 @@ const styles = StyleSheet.create({
     paddingTop: 40
   },
   text: {
-    color: '#fff'
+    color: '#ccc'
   }
 })
